+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Enable passing `string`s to `addPeerToAddressBook`.
 - Use `addPeerToAddressBook` in examples and usage doc.
 - Settle on `js-waku` name across the board.
+- **Breaking**: `RelayDefaultTopic` renamed to `DefaultPubsubTopic`.
 
 ## [0.1.0] - 2021-05-12